+++ src/projects/forms/frm_build.py
@@ -20,7 +20,7 @@ FRAME_TITLE = "Build package"
 
 class BuildFrame:
     def __init__(self, parent, project, git_commit: bool) -> None:
-        self.root = tk.Toplevel(parent)
+        self.root = tk.Toplevel(parent.root)
         self.project = project
         self.git_commit = git_commit