+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.23 - 27 August 2026
 1. Create PARSER_ARGS tuple for module caller
+2. Fix builder call bug
 
 ## Version 1.0.22 - 26 August 2026
 1. Use ModuleCallerBase from psiutils